+++ src/main/java/it/unisa/c02/rently/rently_application/data/dao/GestioneChatDAO.java
@@ -12,5 +12,5 @@ public interface GestioneChatDAO extends JpaRepository<Messaggio, Long> {
 
     @Query("SELECT m FROM Messaggio m WHERE m.mittente = (?1) AND m.destinatario = (?2) UNION" +
             " SELECT m FROM Messaggio m WHERE m.mittente = (?2) AND m.destinatario = (?1)")
-    List<Messaggio> getChat(int id1, int id2);
+    List<Messaggio> getChat(long id1, long id2);
 }
